+++ src/main/java/net/panther/endersteel/enchantment/EndersEdgeEnchantment.java
@@ -1,0 +1,33 @@
+package net.panther.endersteel.enchantment;
+
+import net.minecraft.enchantment.Enchantment;
+import net.minecraft.enchantment.EnchantmentTarget;
+import net.minecraft.entity.EquipmentSlot;
+import net.minecraft.item.ItemStack;
+import net.panther.endersteel.item.custom.EnderSteelSwordItem;
+
+public class EndersEdgeEnchantment extends Enchantment {
+    public EndersEdgeEnchantment() {
+        super(Rarity.RARE, EnchantmentTarget.WEAPON, new EquipmentSlot[]{EquipmentSlot.MAINHAND});
+    }
+
+    @Override
+    public int getMinPower(int level) {
+        return 15;
+    }
+
+    @Override
+    public int getMaxPower(int level) {
+        return 50;
+    }
+
+    @Override
+    public int getMaxLevel() {
+        return 1;
+    }
+
+    @Override
+    public boolean isAcceptableItem(ItemStack stack) {
+        return stack.getItem() instanceof EnderSteelSwordItem;
+    }
+}