+++ src/main/java/net/panther/endersteel/enchantment/ModEnchantments.java
@@ -7,14 +7,19 @@ import net.minecraft.util.Identifier;
 import net.panther.endersteel.EnderSteel;
 
 public class ModEnchantments {
-    public static final Enchantment GAZING_VOID = register("gazing_void",
-            new GazingVoidEnchantment());
+    public static final Enchantment GAZING_VOID = Registry.register(
+            Registries.ENCHANTMENT,
+            new Identifier(EnderSteel.MOD_ID, "gazing_void"),
+            new GazingVoidEnchantment()
+    );
 
-    private static Enchantment register(String name, Enchantment enchantment) {
-        return Registry.register(Registries.ENCHANTMENT, new Identifier(EnderSteel.MOD_ID, name), enchantment);
-    }
+    public static final Enchantment ENDERS_EDGE = Registry.register(
+            Registries.ENCHANTMENT,
+            new Identifier(EnderSteel.MOD_ID, "enders_edge"),
+            new EndersEdgeEnchantment()
+    );
 
     public static void registerModEnchantments() {
-        EnderSteel.LOGGER.info("Registering Mod Enchantments for " + EnderSteel.MOD_ID);
+        EnderSteel.LOGGER.info("Registering Enchantments for " + EnderSteel.MOD_ID);
     }
 }